feature: разбить код на модули
