+++ index.html
@@ -98,6 +98,6 @@
       </article>
     </li>
   </template>
-  <script src="./scripts/script.js"></script>
+  <script src="./scripts/script.js" type="module"></script>
 </body>
 </html>
