+++ scripts/functions.js
@@ -1,0 +1,9 @@
+function openPopup(popupElement) {
+  popupElement.classList.add('popup_opened');
+}
+
+function closePopup(popupElement) {
+  popupElement.classList.remove('popup_opened');
+}
+
+export {openPopup, closePopup};
