+++ scripts/initialCardsLoad.js
@@ -1,0 +1,70 @@
+import { openPopup } from "./functions.js";
+import { photoView, photo, photoTitle } from "./photoModal.js";
+import { places } from "./placesModal.js";
+
+/**
+ * Дефолтный набор карточек
+ */
+export const initialCards = [
+  {
+    name: 'Элиста, Калмыкия',
+    link: './images/kalmikia.jpeg'
+  },
+  {
+    name: 'Алтай',
+    link: './images/altai.jpeg'
+  },
+  {
+    name: 'Чиркейская ГЭС',
+    link: './images/dagestan.jpeg'
+  },
+  {
+    name: 'Домбай',
+    link: './images/dombai.jpg'
+  },
+  {
+    name: 'Эльбрус',
+    link: './images/elbrus.jpg'
+  },
+  {
+    name: 'Карачаевск',
+    link: './images/karachaesk.jpeg'
+  }
+];
+
+
+
+/**
+ * Функция добавления дефолтных карточек + удаление и лайки
+ */
+export function createCard(placeName, placeLink) {
+  const cardTemplate = document.querySelector('#place-card').content;
+  const placeCard = cardTemplate.querySelector('.place').cloneNode(true);
+  const photoCard = placeCard.querySelector('.place__image');
+
+  placeCard.querySelector('.place__title').textContent = placeName;
+  photoCard.src = placeLink;
+  photoCard.alt = placeName;
+
+  photoCard.addEventListener('click', () => {
+    openPopup(photoView);
+    photo.src = placeLink;
+    photo.alt = placeName;
+    photoTitle.textContent = placeName;
+  })
+
+  placeCard.querySelector('.place__button').addEventListener('click', (e) => {
+    e.target.classList.toggle('place__button_active');
+  })
+
+  placeCard.querySelector('.place__delete').addEventListener('click', (e) => {
+    e.target.closest('.place').remove();
+  })
+
+  return placeCard;
+}
+
+export const addInitialCards = (placeName, placeLink) => {
+  const initialPlaceCards = createCard(placeName, placeLink);
+  places.prepend(initialPlaceCards);
+}
