+++ scripts/photoModal.js
@@ -1,0 +1,11 @@
+const photoView = document.querySelector('#photo-view'); //модальное окно, при нажатии на изображение
+const photo = document.querySelector('.photo');
+const photoTitle = document.querySelector('.photo-title');
+const photoContainer = document.querySelector('.popup__photo-container');
+
+export{
+  photoView,
+  photo,
+  photoTitle,
+  photoContainer
+}
