+++ scripts/validationFormFuncs.js
@@ -1,0 +1,58 @@
+function showInputError(form, formInput, errorMessage) {
+  const errorElement = form.querySelector(`.${formInput.id}-error`)
+  formInput.classList.add('popup__input_type_error');
+  errorElement.textContent = errorMessage;
+  errorElement.classList.add('popup__input-error_active');
+}
+
+function hideInputError(form, formInput) {
+  const errorElement = form.querySelector(`.${formInput.id}-error`)
+  formInput.classList.remove('popup__input_type_error');
+  errorElement.classList.remove('popup__input-error_active');
+  errorElement.textContent = '';
+}
+
+function isValid(form, formInput) {
+  if (!formInput.validity.valid) {
+    showInputError(form, formInput, formInput.validationMessage);
+  } else {
+    hideInputError(form, formInput);
+  }
+}
+
+const hasInvalidInput = (inputList) => {
+  return inputList.some(inputElement => {
+    return !inputElement.validity.valid;
+  })
+}
+
+const toggleButtonState = (inputList, buttonElement) => {
+  if (hasInvalidInput(inputList)) {
+    buttonElement.disabled = true;
+    buttonElement.classList.add('popup__button_inactive');
+  } else {
+    buttonElement.disabled = false;
+    buttonElement.classList.remove('popup__button_inactive');
+  }
+}
+
+const setEventListeners = (form) => {
+  const inputList = Array.from(form.querySelectorAll('.popup__input'));
+  const buttonElement = form.querySelector('.popup__button');
+  toggleButtonState(inputList, buttonElement);
+
+  inputList.forEach(inputElement => {
+    inputElement.addEventListener('input', () => {
+      isValid(form, inputElement);
+      toggleButtonState(inputList, buttonElement);
+    })
+  })
+}
+
+export const enableValidation = () => {
+  const formList = Array.from(document.querySelectorAll('.popup__form'));
+
+  formList.forEach(form => {
+    setEventListeners(form);
+  })
+}
